+++ src/parser/download.rs
@@ -1,4 +1,4 @@
-// SPDX-License-Identifier: GPL-3.0-only
+// SPDX-License-Identifier: MIT
 
 use regex::Regex;
 use scraper::{ElementRef, Html, Selector};